fstests: add results for v6.9-rc4
The crashes have been reported, and so testing for those guests
that crashed stopped testing up to that point for the first run.
The expunges observed on kdevops then are for tests which managed
to happen before the crash on each guest that crashed. For all other
guests its the full set of results.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>